backend/models/PlantDetectCNN/requirements.txt
tensorflow>=2.13.0
flask>=2.3.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pandas>=1.4.0
psutil>=5.9.0
GPUtil>=1.4.0
pytest>=7.0.0
requests>=2.28.0